requirements.txt
fastapi
uvicorn
jinja2
pyyaml
markdown
pymdown-extensions
Pillow
libsass
python-multipart
python-jose[cryptography]
bcrypt
watchdog
aiofiles
beautifulsoup4
unidecode
filelock
psutil
libsass
itsdangerous
